Cut/Paste the 10 slowest integral tests into test_slow_integrals.py (each one currently taking more than 5 seconds for me to run), and fix imports. Made it easier to identify that an attempt to fix an evalf()/assumptions issue caused test_transform to take an exceptionally long time of 1040.507 seconds, for reference the slow tests take around 145s, the fast ones take around 40s now.
